implemented merging logic
